requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
jsonfield==3.1.0
mysqlclient==2.2.4
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
pymongo==4.9.1
pytz==2024.2
rest-framework-simplejwt==0.0.2
sqlparse==0.2.4
tzdata==2024.2
